Split photos into photos & timeline, add photo download test.
