+++ crates/apps/api/src/lib.rs
@@ -1,3 +1,4 @@
+#![deny(clippy::unwrap_used)]
 #![allow(
     clippy::needless_for_each,
     clippy::cognitive_complexity,
