+++ crates/apps/api/src/routes/photos/router.rs
@@ -1,16 +1,12 @@
 use crate::api_state::ApiContext;
-use crate::photos::handlers::{
-    get_color_theme_handler, get_full_item_handler, get_photos_by_month_handler, get_random_photo,
-    get_timeline_ids_handler, get_timeline_ratios_handler,
-};
+
 use axum::{Router, routing::get};
+use crate::photos::handlers::{download_full_file, get_color_theme_handler, get_full_item_handler, get_random_photo};
 
 pub fn photos_protected_router() -> Router<ApiContext> {
     Router::new()
         .route("/photos/random", get(get_random_photo))
         .route("/photos/theme", get(get_color_theme_handler))
-        .route("/photos/timeline/ratios", get(get_timeline_ratios_handler))
-        .route("/photos/timeline/ids", get(get_timeline_ids_handler))
-        .route("/photos/by-month", get(get_photos_by_month_handler))
         .route("/photos/item", get(get_full_item_handler))
+        .route("/photos/download", get(download_full_file))
 }
