+++ crates/apps/api/src/routes/timeline/router.rs
@@ -1,0 +1,10 @@
+use crate::api_state::ApiContext;
+use axum::{Router, routing::get};
+use crate::timeline::handlers::{get_photos_by_month_handler, get_timeline_ids_handler, get_timeline_ratios_handler};
+
+pub fn photos_protected_router() -> Router<ApiContext> {
+    Router::new()
+        .route("/timeline/ratios", get(get_timeline_ratios_handler))
+        .route("/timeline/ids", get(get_timeline_ids_handler))
+        .route("/timeline/by-month", get(get_photos_by_month_handler))
+}
