--- crates/libs/common_services/src/api/download/interfaces.rs
@@ -1,8 +1,0 @@
-use serde::Deserialize;
-use utoipa::ToSchema;
-
-#[derive(Deserialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct DownloadMediaParams {
-    pub path: String,
-}
