+++ crates/libs/common_services/src/api/mod.rs
@@ -1,6 +1,6 @@
 pub mod album;
 pub mod auth;
-pub mod download;
 pub mod onboarding;
 pub mod photos;
 pub mod s2s;
+pub mod timeline;
