+++ crates/libs/common_services/src/api/photos/error.rs
@@ -5,7 +5,7 @@ use axum::response::{IntoResponse, Response};
 use color_eyre::eyre;
 use serde_json::json;
 use thiserror::Error;
-use tracing::error;
+use tracing::{error, warn};
 
 #[derive(Debug, Error)]
 pub enum PhotosError {
@@ -15,11 +15,17 @@ pub enum PhotosError {
     #[error("internal error")]
     Internal(#[from] eyre::Report),
 
-    #[error("Invalid month format. Expected YYYY-MM-DD, but got '{0}'")]
-    InvalidMonthFormat(String),
-
     #[error("Media item not found: {0}")]
     MediaNotFound(String),
+
+    #[error("Invalid media path provided.")]
+    InvalidPath,
+
+    #[error("Permission denied while accessing media file.")]
+    AccessDenied,
+
+    #[error("The requested file is not a supported media type.")]
+    UnsupportedMediaType,
 }
 
 // Renamed for more general use
@@ -27,12 +33,12 @@ fn log_error(error: &PhotosError) {
     match error {
         PhotosError::Database(e) => error!("Database query failed: {}", e),
         PhotosError::Internal(e) => error!("Internal error: {}", e),
-        PhotosError::InvalidMonthFormat(month) => {
-            error!("Invalid month format provided: {}", month);
-        }
         PhotosError::MediaNotFound(id) => {
-            error!("Media item not found: {}", id);
-        }
+            warn!("Media item not found: {}", id);
+        },
+        PhotosError::InvalidPath => warn!("Invalid path provided."),
+        PhotosError::AccessDenied => warn!("access denied"),
+        PhotosError::UnsupportedMediaType => warn!("unsupported media type."),
     }
 }
 
@@ -49,14 +55,13 @@ impl IntoResponse for PhotosError {
                 StatusCode::INTERNAL_SERVER_ERROR,
                 "An unexpected internal error occurred.".to_string(),
             ),
-            Self::InvalidMonthFormat(invalid_month) => (
-                StatusCode::BAD_REQUEST,
-                format!("Invalid month format. Expected YYYY-MM-DD, but got '{invalid_month}'"),
-            ),
             Self::MediaNotFound(media_id) => (
                 StatusCode::NOT_FOUND,
                 format!("Media item not found: {media_id}"),
             ),
+            Self::InvalidPath => (StatusCode::BAD_REQUEST, self.to_string()),
+            Self::AccessDenied => (StatusCode::FORBIDDEN, self.to_string()),
+            Self::UnsupportedMediaType => (StatusCode::UNSUPPORTED_MEDIA_TYPE, self.to_string()),
         };
 
         let body = Json(json!({ "error": error_message }));
