+++ crates/libs/common_services/src/api/photos/service.rs
@@ -1,15 +1,17 @@
 use crate::api::photos::error::PhotosError;
 
 use crate::api::photos::interfaces::RandomPhotoResponse;
-use crate::database::app_user::User;
-use chrono::NaiveDate;
-use common_types::pb::api::{
-    ByMonthResponse, MediaItem, MediaMonth, TimelineMonth, TimelineResponse,
-};
+use crate::database::app_user::{User, UserRole};
 use rand::Rng;
 use sqlx::PgPool;
-use std::collections::HashMap;
-use tracing::warn;
+use std::path::Path;
+use axum::body::Body;
+use color_eyre::Report;
+use http::{header, Response, StatusCode};
+use tokio::fs::File;
+use tokio_util::codec::{BytesCodec, FramedRead};
+use tracing::{debug, warn};
+use app_state::{to_posix_string, IngestSettings, MakeRelativePath};
 
 /// Fetches a random photo with its color theme data for a specific user.
 ///
@@ -76,101 +78,95 @@ pub async fn random_photo(
     Ok(random_data)
 }
 
-/// Fetches a timeline of media item ratios, grouped by month.
+
+/// Securely streams a validated media file to the client after performing authorization checks.
 ///
 /// # Errors
 ///
-/// Returns an error if the database query fails.
-pub async fn get_timeline_ratios(
+/// Returns a `PhotosError` if path validation fails, the user is not authorized,
+/// the file is not found, the media type is unsupported, or if any file system
+/// or response building error occurs.
+pub async fn download_media_file(
+    ingestion: &IngestSettings,
     user: &User,
-    pool: &PgPool,
-) -> Result<TimelineResponse, PhotosError> {
-    let months = sqlx::query_as!(
-        TimelineMonth,
-        r#"
-        SELECT
-            month_id::TEXT as "month_id!",
-            COUNT(*)::INT AS "count!",
-            array_agg(width::real / height::real ORDER BY sort_timestamp DESC) AS "ratios!"
-        FROM media_item
-        WHERE user_id = $1
-          AND deleted = false
-        GROUP BY month_id
-        ORDER BY month_id DESC
-        "#,
-        user.id
-    )
-    .fetch_all(pool)
-    .await?;
+    path: &str,
+) -> Result<Response<Body>, PhotosError> {
+    // --- 1. Security & Path Validation ---
+    let media_dir_canon = ingestion
+        .media_root
+        .canonicalize()
+        .map_err(|e| Report::new(e).wrap_err("Failed to canonicalize media directory"))?;
 
-    Ok(TimelineResponse { months })
-}
+    let file_path = ingestion.media_root.join(path);
+    let file_canon = match file_path.canonicalize() {
+        Ok(path) => path,
+        Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
+            debug!("File not found at path: {}", file_path.display());
+            return Err(PhotosError::MediaNotFound(to_posix_string(&file_path)));
+        }
+        Err(e) => {
+            return Err(Report::new(e)
+                .wrap_err("Failed to canonicalize path")
+                .into());
+        }
+    };
 
-/// Fetches a timeline of media item ids.
-///
-/// # Errors
-///
-/// Returns an error if the database query fails.
-pub async fn get_timeline_ids(user: &User, pool: &PgPool) -> Result<Vec<String>, PhotosError> {
-    let months = sqlx::query_scalar!(
-        r#"
-        SELECT id 
-        FROM media_item 
-        WHERE user_id = $1 AND deleted IS NOT TRUE 
-        ORDER BY sort_timestamp DESC
-        "#,
-        user.id
-    )
-    .fetch_all(pool)
-    .await?;
-
-    Ok(months)
-}
-
-/// Fetches media items for a given list of month IDs, grouped by month.
-///
-/// # Errors
-///
-/// Returns an error if the database query fails.
-pub async fn get_photos_by_month(
-    user: &User,
-    pool: &PgPool,
-    month_ids: &[NaiveDate],
-) -> Result<ByMonthResponse, PhotosError> {
-    let items = sqlx::query_as!(
-        MediaItem,
-        r#"
-        SELECT
-            id as "id!",
-            is_video as "is_video!",
-            use_panorama_viewer as "is_panorama!",
-            duration_ms::INT,
-            taken_at_local::TEXT as "timestamp!"
-        FROM
-            media_item
-        WHERE
-            user_id = $1
-            AND deleted = false
-            AND month_id = ANY($2)
-        ORDER BY
-            sort_timestamp DESC
-        "#,
-        user.id,
-        month_ids,
-    )
-    .fetch_all(pool)
-    .await?;
-
-    let mut months_map: HashMap<String, Vec<MediaItem>> = HashMap::new();
-    for item in items {
-        let month_id = format!("{}-01", &item.timestamp[0..7]);
-        months_map.entry(month_id).or_default().push(item);
+    if !file_canon.starts_with(&media_dir_canon) {
+        warn!("Blocked directory traversal attempt for: {}", path);
+        return Err(PhotosError::InvalidPath);
     }
 
-    let months = months_map
-        .into_iter()
-        .map(|(month_id, items)| MediaMonth { month_id, items })
-        .collect();
+    // --- 2. Authorization ---
+    let relative_path = file_canon.make_relative_canon(&ingestion.media_root_canon)?;
+    if user.role != UserRole::Admin {
+        let Some(user_media_folder) = &user.media_folder else {
+            warn!(
+                "Access denied for user {}: No media folder assigned.",
+                user.id
+            );
+            return Err(PhotosError::AccessDenied);
+        };
+        if !relative_path.starts_with(user_media_folder) {
+            warn!(
+                "Access denied for user {}: Attempted to access path outside their media folder.",
+                user.id
+            );
+            return Err(PhotosError::AccessDenied);
+        }
+    }
 
-    Ok(ByMonthResponse { months })
+    // --- 3. Media Type Validation ---
+    if !ingestion.is_media_file(&file_canon) {
+        warn!("Unsupported media type requested: {}", file_canon.display());
+        return Err(PhotosError::UnsupportedMediaType);
+    }
+
+    // --- 4. File Access ---
+    let file = match File::open(&file_canon).await {
+        Ok(file) => file,
+        Err(e) => match e.kind() {
+            std::io::ErrorKind::NotFound => Err(PhotosError::MediaNotFound(to_posix_string(&file_canon))),
+            std::io::ErrorKind::PermissionDenied => Err(PhotosError::AccessDenied),
+            _ => Err(Report::new(e).wrap_err("Failed to open media file").into()),
+        }?,
+    };
+
+    // --- 5. Build the Streaming Response ---
+    let stream = FramedRead::new(file, BytesCodec::new());
+    let body = Body::from_stream(stream);
+    let mime_type = mime_guess::from_path(&file_canon).first_or_octet_stream();
+    let filename = Path::new(path)
+        .file_name()
+        .and_then(|n| n.to_str())
+        .unwrap_or("mediafile");
+    let disposition = format!("inline; filename=\"{filename}\"");
+    let disposition_header = header::HeaderValue::from_str(&disposition)
+        .unwrap_or(header::HeaderValue::from_static("inline"));
+
+    Ok(Response::builder()
+        .status(StatusCode::OK)
+        .header(header::CONTENT_TYPE, mime_type.as_ref())
+        .header(header::CONTENT_DISPOSITION, disposition_header)
+        .body(body)
+        .map_err(|e| Report::new(e).wrap_err("Failed to build response"))?)
 }
