+++ crates/libs/common_services/src/api/timeline/interfaces.rs
@@ -1,0 +1,9 @@
+use serde::Deserialize;
+use utoipa::{IntoParams, ToSchema};
+
+#[derive(Deserialize, IntoParams, ToSchema, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct GetMediaByMonthParams {
+    /// "YYYY-MM-DD" strings.
+    pub months: String,
+}
