+++ crates/libs/common_services/src/lib.rs
@@ -1,3 +1,4 @@
+#![deny(clippy::unwrap_used)]
 #![allow(
     clippy::missing_errors_doc,
     clippy::missing_panics_doc,
