+++ crates/libs/common_types/src/ml_analysis.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq)]
+#[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Default)]
 pub struct PyOCRData {
     pub has_legible_text: bool,
     pub ocr_text: Option<String>,
