+++ crates/test_integration/src/tests/mod.rs
@@ -2,4 +2,4 @@ pub mod test_album;
 pub mod test_auth;
 pub mod test_onboarding;
 pub mod test_root;
-pub mod test_download;
+pub mod test_photos;
