--- crates/test_integration/src/tests/test_download.rs
@@ -1,1 +1,0 @@
-// todo test download handler here.
